Refactor dynamical systems and integration protocols
- Moved `_DynamicalSystemProtocol` and `_HamiltonianSystemProtocol` to a new `protocols` module for better organization and clarity.
- Updated import statements across various files to reflect the new location of the protocols.
- Enhanced the integration methods to utilize the updated protocol imports, ensuring consistency and improved maintainability.
- Removed unused protocol definitions and streamlined the codebase for better performance.
