+++ src/hiten/algorithms/poincare/centermanifold/backend.py
@@ -18,8 +18,8 @@ from numba import njit, prange
 from scipy.optimize import root_scalar
 
 from hiten.algorithms.dynamics.hamiltonian import (_eval_dH_dP, _eval_dH_dQ,
-                                                   _hamiltonian_rhs,
-                                                   _HamiltonianSystemProtocol)
+                                                   _hamiltonian_rhs)
+from hiten.algorithms.dynamics.protocols import _HamiltonianSystemProtocol
 from hiten.algorithms.integrators.rk import (RK4_A, RK4_B, RK4_C, RK6_A, RK6_B,
                                              RK6_C, RK8_A, RK8_B, RK8_C)
 from hiten.algorithms.integrators.symplectic import (N_SYMPLECTIC_DOF,
@@ -393,7 +393,7 @@ class _CenterManifoldBackend(_ReturnMapBackend):
 
     Parameters
     ----------
-    dynsys : :class:`~hiten.algorithms.dynamics.hamiltonian._HamiltonianSystemProtocol`
+    dynsys : :class:`~hiten.algorithms.dynamics.protocols._HamiltonianSystemProtocol`
         Hamiltonian system providing polynomial representation and CLMO tables.
     surface : :class:`~hiten.algorithms.poincare.core.events._SurfaceEvent`
         Poincare section surface definition.
